refactor: reuse registration success ui
