+++ src/components/features/registration/ProviderRegistrationSuccess/index.ts
@@ -1,2 +1,1 @@
-export * from './ProviderRegistrationFlow';
 export * from './ProviderRegistrationSuccess';
