+++ src/lib/sitemap/providers/coaches/index.ts
@@ -1,0 +1,4 @@
+export const COACH_PATHS = {
+    REGISTER: '/providers/coach/register',
+    REGISTER_SUCCESS: '/providers/coach/register/success',
+} as const;
