+++ src/lib/sitemap/providers/index.ts
@@ -1,4 +1,7 @@
+import { COACH_PATHS } from './coaches';
+import { THERAPIST_PATHS } from './therapists';
+
 export const PROVIDERS_PATHS = {
-    REGISTER: '/providers/register',
-    REGISTER_SUCCESS: '/providers/register/success',
+    COACH: COACH_PATHS,
+    THERAPIST: THERAPIST_PATHS,
 } as const;
