+++ src/lib/sitemap/providers/therapists/index.ts
@@ -1,0 +1,4 @@
+export const THERAPIST_PATHS = {
+    REGISTER: '/providers/therapist/register',
+    REGISTER_SUCCESS: '/providers/therapist/register/success',
+} as const;
